+++ Libplanet.Action.Tests/ActionContextTest.cs
@@ -12,12 +12,15 @@ namespace Libplanet.Action.Tests
         private readonly System.Random _random;
         private readonly Address _address;
         private readonly TxId _txid;
+        private readonly ISystemAccounts _systemAccounts;
 
         public ActionContextTest()
         {
             _random = new System.Random();
             _address = _random.NextAddress();
             _txid = _random.NextTxId();
+            _systemAccounts = new SystemAccounts(
+                new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount), null);
         }
 
         [Fact]
@@ -34,6 +37,7 @@ namespace Libplanet.Action.Tests
                     signer: _address,
                     txid: _txid,
                     miner: _address,
+                    systemAccounts: _systemAccounts,
                     blockIndex: 1,
                     blockProtocolVersion: Block.CurrentProtocolVersion,
                     previousState: new World(new MockWorldState()),
@@ -52,6 +56,7 @@ namespace Libplanet.Action.Tests
                 signer: _address,
                 txid: _txid,
                 miner: _address,
+                systemAccounts: _systemAccounts,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
                 previousState: new World(new MockWorldState()),
@@ -63,6 +68,7 @@ namespace Libplanet.Action.Tests
                 signer: _address,
                 txid: _txid,
                 miner: _address,
+                systemAccounts: _systemAccounts,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
                 previousState: new World(new MockWorldState()),
@@ -74,6 +80,7 @@ namespace Libplanet.Action.Tests
                 signer: _address,
                 txid: _txid,
                 miner: _address,
+                systemAccounts: _systemAccounts,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
                 previousState: new World(new MockWorldState()),
@@ -113,6 +120,7 @@ namespace Libplanet.Action.Tests
                     signer: _address,
                     txid: _txid,
                     miner: _address,
+                    systemAccounts: _systemAccounts,
                     blockIndex: 1,
                     blockProtocolVersion: Block.CurrentProtocolVersion,
                     previousState: new World(new MockWorldState()),