+++ Libplanet.Action/ActionContext.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics.Contracts;
 using System.Threading;
 using Libplanet.Action.State;
 using Libplanet.Crypto;
@@ -17,6 +16,7 @@ namespace Libplanet.Action
             Address signer,
             TxId? txid,
             Address miner,
+            ISystemAccounts systemAccounts,
             long blockIndex,
             int blockProtocolVersion,
             IWorld previousState,
@@ -27,6 +27,7 @@ namespace Libplanet.Action
             Signer = signer;
             TxId = txid;
             Miner = miner;
+            SystemAccounts = systemAccounts;
             BlockIndex = blockIndex;
             BlockProtocolVersion = blockProtocolVersion;
             Rehearsal = rehearsal;
@@ -45,6 +46,9 @@ namespace Libplanet.Action
 
         /// <inheritdoc cref="IActionContext.Miner"/>
         public Address Miner { get; }
+
+        /// <inheritdoc cref="IActionContext.SystemAccounts"/>
+        public ISystemAccounts SystemAccounts { get; }
 
         /// <inheritdoc cref="IActionContext.BlockIndex"/>
         public long BlockIndex { get; }