+++ Libplanet.Action/ActionEvaluator.cs
@@ -24,6 +24,7 @@ namespace Libplanet.Action
     public class ActionEvaluator : IActionEvaluator
     {
         private readonly ILogger _logger;
+        private readonly ISystemAccountsGetter _systemAccountsGetter;
         private readonly PolicyBlockActionGetter _policyBlockActionGetter;
         private readonly IStateStore _stateStore;
         private readonly IActionLoader _actionLoader;
@@ -33,17 +34,21 @@ namespace Libplanet.Action
         /// </summary>
         /// <param name="policyBlockActionGetter">A delegator to get policy block action to evaluate
         /// at the end for each <see cref="IPreEvaluationBlock"/> that gets evaluated.</param>
+        /// <param name="systemAccountsGetter">The <see cref="ISystemAccountsGetter"/> to determine
+        /// <see cref="ISystemAccounts"/> with <see cref="IPreEvaluationBlockHeader"/>.</param>
         /// <param name="stateStore">The <see cref="IStateStore"/> to use to retrieve
         /// the states for a provided <see cref="HashDigest{SHA256}"/>.</param>
         /// <param name="actionTypeLoader"> A <see cref="IActionLoader"/> implementation using
         /// action type lookup.</param>
         public ActionEvaluator(
+            ISystemAccountsGetter systemAccountsGetter,
             PolicyBlockActionGetter policyBlockActionGetter,
             IStateStore stateStore,
             IActionLoader actionTypeLoader)
         {
             _logger = Log.ForContext<ActionEvaluator>()
                 .ForContext("Source", nameof(ActionEvaluator));
+            _systemAccountsGetter = systemAccountsGetter;
             _policyBlockActionGetter = policyBlockActionGetter;
             _stateStore = stateStore;
             _actionLoader = actionTypeLoader;
@@ -130,6 +135,8 @@ namespace Libplanet.Action
         /// </summary>
         /// <param name="blockHeader">The <see cref="IPreEvaluationBlockHeader"/> of
         /// the <see cref="Block"/> that <paramref name="actions"/> belong to.</param>
+        /// <param name="systemAccountsGetter">The <see cref="ISystemAccountsGetter"/> to determine
+        /// <see cref="ISystemAccounts"/> with <see cref="IPreEvaluationBlockHeader"/>.</param>
         /// <param name="tx">The <see cref="Transaction"/> that <paramref name="actions"/>
         /// belong to.  This should be <see langword="null"/> if <paramref name="actions"/>
         /// do not belong to a <see cref="Transaction"/>, i.e.
@@ -167,6 +174,7 @@ namespace Libplanet.Action
         [Pure]
         internal static IEnumerable<ActionEvaluation> EvaluateActions(
             IPreEvaluationBlockHeader blockHeader,
+            ISystemAccountsGetter systemAccountsGetter,
             ITransaction? tx,
             IWorld previousState,
             IImmutableList<IAction> actions,
@@ -181,6 +189,7 @@ namespace Libplanet.Action
                     signer: tx?.Signer ?? blockHeader.Miner,
                     txid: tx?.Id ?? null,
                     miner: blockHeader.Miner,
+                    systemAccounts: new SystemAccounts(systemAccountsGetter, blockHeader),
                     blockIndex: blockHeader.Index,
                     blockProtocolVersion: blockHeader.ProtocolVersion,
                     previousState: prevState,
@@ -241,6 +250,7 @@ namespace Libplanet.Action
                     signer: inputContext.Signer,
                     txid: inputContext.TxId,
                     miner: inputContext.Miner,
+                    systemAccounts: inputContext.SystemAccounts,
                     blockIndex: inputContext.BlockIndex,
                     blockProtocolVersion: inputContext.BlockProtocolVersion,
                     previousState: newPrevState,
@@ -433,6 +443,7 @@ namespace Libplanet.Action
                 ImmutableList.CreateRange(LoadActions(blockHeader.Index, tx));
             return EvaluateActions(
                 blockHeader: blockHeader,
+                systemAccountsGetter: _systemAccountsGetter,
                 tx: tx,
                 previousState: previousState,
                 actions: actions,
@@ -470,6 +481,7 @@ namespace Libplanet.Action
 
             return EvaluateActions(
                 blockHeader: blockHeader,
+                systemAccountsGetter: _systemAccountsGetter,
                 tx: null,
                 previousState: previousState,
                 actions: new[] { policyBlockAction }.ToImmutableList()).Single();
@@ -569,7 +581,7 @@ namespace Libplanet.Action
             stopwatch.Start();
 
             var totalDelta = evaluation.OutputState.GetAccount(
-                ReservedAddresses.LegacyAccount).Delta.ToRawDelta();
+                ReservedAddresses.DefaultAccount).Delta.ToRawDelta();
 
             int setCount = 0;
 
@@ -644,7 +656,7 @@ namespace Libplanet.Action
         {
             Stopwatch stopwatch = new Stopwatch();
             worldTrie = _stateStore.GetStateRoot(null).Set(
-                KeyConverters.ToStateKey(ReservedAddresses.LegacyAccount),
+                KeyConverters.ToStateKey(ReservedAddresses.DefaultAccount),
                 new Binary(worldTrie.Hash.ByteArray));
 
             _logger