+++ Libplanet.Explorer.Tests/Queries/StateQueryTest.cs
@@ -29,7 +29,8 @@ public class StateQueryTest
     public async Task WorldStates()
     {
         (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy()
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -57,7 +58,8 @@ public class StateQueryTest
     public async Task AccountStates()
     {
         (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy()
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -84,7 +86,8 @@ public class StateQueryTest
     public async Task State()
     {
         (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy()
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -110,7 +113,7 @@ public class StateQueryTest
     {
         (IBlockChainStates, IBlockPolicy) source = (
             new MockChainStates(),
-            new BlockPolicy()
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -155,7 +158,8 @@ public class StateQueryTest
         var legacyToken = Currency.Legacy("LEG", 0, null);
 #pragma warning restore CS0618
         (IBlockChainStates, IBlockPolicy) source = (
-           new MockChainStates(), new BlockPolicy());
+           new MockChainStates(),
+           new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount)));
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
             totalSupply(
@@ -209,7 +213,7 @@ public class StateQueryTest
     {
         (IBlockChainStates, IBlockPolicy) source = (
            new MockChainStates(),
-           new BlockPolicy()
+           new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
        );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -238,7 +242,8 @@ public class StateQueryTest
     public async Task ThrowExecutionErrorIfViolateMutualExclusive()
     {
         (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy()
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -259,7 +264,8 @@ public class StateQueryTest
     {
         var currency = Currency.Uncapped("ABC", 2, minters: null);
         (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy()
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -285,7 +291,7 @@ public class StateQueryTest
     {
         (IBlockChainStates, IBlockPolicy) source = (
             new MockChainStates(),
-            new BlockPolicy()
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -330,7 +336,8 @@ public class StateQueryTest
          var legacyToken = Currency.Legacy("LEG", 0, null);
 #pragma warning restore CS0618
          (IBlockChainStates, IBlockPolicy) source = (
-            new MockChainStates(), new BlockPolicy());
+            new MockChainStates(),
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount)));
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {
             totalSupply(
@@ -385,7 +392,7 @@ public class StateQueryTest
     {
          (IBlockChainStates, IBlockPolicy) source = (
             new MockChainStates(),
-            new BlockPolicy()
+            new BlockPolicy(new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount))
         );
         ExecutionResult result = await ExecuteQueryAsync<StateQuery>(@"
         {