+++ Libplanet.Extensions.Cocona/BlockPolicyParams.cs
@@ -1,7 +1,6 @@
 namespace Libplanet.Extensions.Cocona;
 
 using System;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
@@ -74,6 +73,9 @@ public class BlockPolicyParams : ICommandParameterSet
     public IAction? GetBlockAction() =>
         GetBlockAction(LoadAssemblies());
 
+    public ISystemAccountsGetter? GetSystemAccountsGetter() =>
+        GetSystemAccountsGetter(LoadAssemblies());
+
     [SuppressMessage(
         "Major Code Smell",
         "S3011:Reflection should not be used to increase accessibility of classes, methods, " +
@@ -144,5 +146,19 @@ public class BlockPolicyParams : ICommandParameterSet
             .GetType()
             .GetProperty(nameof(IBlockPolicy.BlockAction));
         return (IAction?)prop!.GetValue(policy);
+    }
+
+    internal ISystemAccountsGetter? GetSystemAccountsGetter(Assembly[] assemblies)
+    {
+        object? policy = GetBlockPolicy(assemblies);
+        if (policy is null)
+        {
+            return null;
+        }
+
+        PropertyInfo? prop = policy
+            .GetType()
+            .GetProperty(nameof(IBlockPolicy.SystemAccountsGetter));
+        return (ISystemAccountsGetter?)prop!.GetValue(policy);
     }
 }