+++ Libplanet.Tests/Blockchain/Renderers/AnonymousActionRendererTest.cs
@@ -16,6 +16,9 @@ namespace Libplanet.Tests.Blockchain.Renderers
     {
         private static IValue _action = new DumbAction().PlainValue;
 
+        private static ISystemAccounts _systemAccounts = new SystemAccounts(
+            new SystemAccountsGetter(_ => ReservedAddresses.DefaultAccount), null);
+
         private static IWorld _world = new World(new MockWorldState());
 
         private static ICommittedActionContext _actionContext =
@@ -23,8 +26,9 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 default,
                 default,
                 default,
-                Block.CurrentProtocolVersion,
+                _systemAccounts,
                 default,
+                Block.CurrentProtocolVersion,
                 _world,
                 default,
                 0));